// ui-inbox/src/settings/integrations/types.ts
export interface IBookingData {
  name?: string;
  image?: any;
  description?: string;
  userFilters?: string[];
  productCategoryId?: string;
  style?: IStyle;
  mainProductCategory?: IProductCategory;
  navigationText?: string;
  bookingFormText?: string;

  viewCount?: number;
  productFieldIds?: string[];
}
// ...
export interface IMessengerData {
  botEndpointUrl?: string;
  botShowInitialMessage?: boolean;
  skillData?: ISkillData;
  messages?: IMessages;
  notifyCustomer?: boolean;
  supporterIds?: string[];
  availabilityMethod?: string;
  isOnline?: boolean;
  timezone?: string;
  responseRate?: string;
  showTimezone?: boolean;
  requireAuth?: boolean;
  showChat?: boolean;
  showLauncher?: boolean;
  forceLogoutWhenResolve?: boolean;
  showVideoCallRequest?: boolean;
  onlineHours?: IOnlineHour[];
  hideWhenOffline?: boolean;
  links?: ILink;
  externalLinks?: IExternalLink[];
}

export interface IUiOptions {
  color?: string;
  textColor?: string;
  wallpaper?: string;
  logo?: string;
  logoPreviewUrl?: string;
}

export interface ITopic {
  topicId: string;
}

export interface ITopicMessengerApp {
  credentials: ITopic;
}

export interface IWebsite {
  url: string;
  buttonText: string;
  description: string;
}

export interface IWebsiteMessengerApp {
  credentials: IWebsite;
}

export interface ILead {
  formCode: string;
}

export interface ILeadMessengerApp {
  credentials: ILead;
}

interface IIntegartionHealthStatus {
  status: string;
  error: string;
}

export interface IIntegration {
  _id: string;
  kind: string;
  logo?: string;
  name: string;
  brandId?: string;
  code: string;
  formId: string;
  languageCode?: string;
  createUrl: string;
  messengerData?: IMessengerData;
  form: IForm;
  uiOptions?: IUiOptions;
  leadData: ILeadData;
  brand: IBrand;
  channels: any[];
  isActive?: boolean;
  isConnected?: boolean;
  healthStatus?: IIntegartionHealthStatus;
  webhookData?: IWebhookData;
  leadMessengerApps?: ILeadMessengerApp[];
  websiteMessengerApps?: IWebsiteMessengerApp[];
  knowledgeBaseMessengerApps?: ITopicMessengerApp[];
  bookingData?: IBookingData;
  visibility?: string;
  departmentIds?: string[];
  details?: any;
}

export type QueryVariables = {
  page?: number;
  perPage?: number;
  searchValue?: string;
  excludeChannelId?: string;
};

export type IntegrationsQueryResponse = {
  integrations: IIntegration[];
  loading: boolean;
  refetch: (variables?: QueryVariables) => Promise<any>;
};